+++ src/main/java/ladder/domain/ladder/LadderHeight.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import ladder.exception.LadderHeightBoundException;
